add detailed info page
