+++ src/components/Layout.tsx
@@ -1,4 +1,4 @@
-import Header from '../header/Header'
+import Header from './Header'
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
   return (
