+++ src/components/Register.tsx
@@ -1,8 +1,8 @@
 import { useDispatch } from 'react-redux'
 import { useNavigate } from 'react-router-dom'
 import { getAuth, createUserWithEmailAndPassword } from 'firebase/auth'
-import { Form } from '../forms/Form'
-import { setUser } from '../../store/slices/userSlice'
+import { Form } from './Form'
+import { setUser } from '../store/slices/userSlice'
 
 const Register = () => {
   const dispatch = useDispatch()
