+++ src/pages/FavoritesPage.tsx
