+++ src/pages/LoginPage.tsx
@@ -1,4 +1,4 @@
-import { Login } from '../../components/login/Login'
+import { Login } from '../components/Login'
 import React from 'react'
 import { Link } from 'react-router-dom'
